Fix getNodeComment import issue by moving to utils.ts
