+++ packages/plugins/other/visitor-plugin-common/src/index.ts
@@ -18,4 +18,3 @@ export * from './types.js';
 export * from './utils.js';
 export * from './variables-to-object.js';
 export * from './convert-schema-enum-to-declaration-block-string.js';
-export * from './get-node-comment.js';
