+++ packages/plugins/other/visitor-plugin-common/src/utils.ts
@@ -22,6 +22,9 @@ import {
   StringValueNode,
   TypeNode,
   DirectiveNode,
+  FieldDefinitionNode,
+  EnumValueDefinitionNode,
+  InputValueDefinitionNode,
 } from 'graphql';
 import { RawConfig } from './base-visitor.js';
 import { parseMapper } from './mappers.js';
@@ -666,4 +669,28 @@ export const getFieldNames = ({
     }
   }
   return fieldNames;
+};
+
+export const getNodeComment = (
+  node: FieldDefinitionNode | EnumValueDefinitionNode | InputValueDefinitionNode
+): string => {
+  let commentText = node.description?.value;
+  const deprecationDirective = node.directives.find(v => v.name.value === 'deprecated');
+  if (deprecationDirective) {
+    const deprecationReason = getDeprecationReason(deprecationDirective);
+    commentText = `${commentText ? `${commentText}\n` : ''}@deprecated ${deprecationReason}`;
+  }
+  const comment = transformComment(commentText, 1);
+  return comment;
+};
+
+const getDeprecationReason = (directive: DirectiveNode): string | void => {
+  if (directive.name.value === 'deprecated') {
+    let reason = 'Field no longer supported';
+    const deprecatedReason = directive.arguments[0];
+    if (deprecatedReason && deprecatedReason.value.kind === Kind.STRING) {
+      reason = deprecatedReason.value.value;
+    }
+    return reason;
+  }
 };
